fix(core): extract episode verify checks, add rollback test for supersession
