+++ packages/engram-core/src/format/verify.ts
@@ -5,13 +5,12 @@
  * valid = true means no error-severity violations (warnings are acceptable).
  */
 
-import {
-  ENTITY_TYPES,
-  EPISODE_SOURCE_TYPES,
-  INGESTION_SOURCE_TYPES,
-  RELATION_TYPES,
-} from "../vocab/index.js";
 import type { EngramGraph } from "./graph.js";
+import {
+  checkEpisodeDanglingSupersededBy,
+  checkEpisodeFanIn,
+  checkVocab,
+} from "./verify-episodes.js";
 import { FORMAT_VERSION, MIN_READABLE_VERSION } from "./version.js";
 
 function compareSemver(a: string, b: string): number {
@@ -363,46 +362,6 @@ function checkProjectionDependencyCycles(graph: EngramGraph): Violation[] {
   return violations;
 }
 
-function checkEpisodeFanIn(graph: EngramGraph): Violation[] {
-  // Each episode may have at most one successor (fan-in uniqueness).
-  // i.e., no two episodes should share the same superseded_by value.
-  const rows = graph.db
-    .query<{ superseded_by: string; cnt: number }, []>(
-      `SELECT superseded_by, COUNT(*) AS cnt
-       FROM episodes
-       WHERE superseded_by IS NOT NULL
-       GROUP BY superseded_by
-       HAVING cnt > 1`,
-    )
-    .all();
-
-  return rows.map((row) => ({
-    check: "checkEpisodeFanIn",
-    entity_or_edge_id: row.superseded_by,
-    message: `Episode '${row.superseded_by}' is referenced as superseded_by by ${row.cnt} episodes (fan-in violation — each episode may have at most one successor)`,
-    severity: "error" as ViolationSeverity,
-  }));
-}
-
-function checkEpisodeDanglingSupersededBy(graph: EngramGraph): Violation[] {
-  // superseded_by must point to a valid existing episode id.
-  const rows = graph.db
-    .query<{ id: string; superseded_by: string }, []>(
-      `SELECT id, superseded_by
-       FROM episodes
-       WHERE superseded_by IS NOT NULL
-         AND superseded_by NOT IN (SELECT id FROM episodes)`,
-    )
-    .all();
-
-  return rows.map((row) => ({
-    check: "checkEpisodeDanglingSupersededBy",
-    entity_or_edge_id: row.id,
-    message: `Episode '${row.id}' has superseded_by '${row.superseded_by}' which does not exist`,
-    severity: "error" as ViolationSeverity,
-  }));
-}
-
 function checkActiveEdgeOverlaps(graph: EngramGraph): Violation[] {
   // Find pairs of active edges (no invalidated_at) sharing the same
   // (source_id, target_id, relation_type, edge_kind) with overlapping validity windows.
@@ -433,87 +392,6 @@ function checkActiveEdgeOverlaps(graph: EngramGraph): Violation[] {
     message: `Active edges '${row.id_a}' and '${row.id_b}' share the same (source, target, relation, kind) with overlapping validity windows`,
     severity: "warning" as ViolationSeverity,
   }));
-}
-
-const KNOWN_ENTITY_TYPES = new Set(Object.values(ENTITY_TYPES));
-const KNOWN_EPISODE_SOURCE_TYPES = new Set(Object.values(EPISODE_SOURCE_TYPES));
-const KNOWN_INGESTION_SOURCE_TYPES = new Set(
-  Object.values(INGESTION_SOURCE_TYPES),
-);
-const KNOWN_RELATION_TYPES = new Set(Object.values(RELATION_TYPES));
-
-function checkVocab(graph: EngramGraph): Violation[] {
-  const violations: Violation[] = [];
-
-  // entities.entity_type
-  const entityRows = graph.db
-    .query<{ id: string; entity_type: string }, []>(
-      "SELECT id, entity_type FROM entities",
-    )
-    .all();
-  for (const row of entityRows) {
-    if (!KNOWN_ENTITY_TYPES.has(row.entity_type)) {
-      violations.push({
-        check: "checkVocab",
-        entity_or_edge_id: row.id,
-        message: `Entity '${row.id}' has unknown entity_type '${row.entity_type}' (not in ENTITY_TYPES registry)`,
-        severity: "warning",
-      });
-    }
-  }
-
-  // episodes.source_type
-  const episodeRows = graph.db
-    .query<{ id: string; source_type: string }, []>(
-      "SELECT id, source_type FROM episodes WHERE status != 'redacted'",
-    )
-    .all();
-  for (const row of episodeRows) {
-    if (!KNOWN_EPISODE_SOURCE_TYPES.has(row.source_type)) {
-      violations.push({
-        check: "checkVocab",
-        entity_or_edge_id: row.id,
-        message: `Episode '${row.id}' has unknown source_type '${row.source_type}' (not in EPISODE_SOURCE_TYPES registry)`,
-        severity: "warning",
-      });
-    }
-  }
-
-  // ingestion_runs.source_type
-  const runRows = graph.db
-    .query<{ id: string; source_type: string }, []>(
-      "SELECT id, source_type FROM ingestion_runs",
-    )
-    .all();
-  for (const row of runRows) {
-    if (!KNOWN_INGESTION_SOURCE_TYPES.has(row.source_type)) {
-      violations.push({
-        check: "checkVocab",
-        entity_or_edge_id: row.id,
-        message: `IngestionRun '${row.id}' has unknown source_type '${row.source_type}' (not in INGESTION_SOURCE_TYPES registry)`,
-        severity: "warning",
-      });
-    }
-  }
-
-  // edges.relation_type
-  const edgeRows = graph.db
-    .query<{ id: string; relation_type: string }, []>(
-      "SELECT id, relation_type FROM edges WHERE invalidated_at IS NULL",
-    )
-    .all();
-  for (const row of edgeRows) {
-    if (!KNOWN_RELATION_TYPES.has(row.relation_type)) {
-      violations.push({
-        check: "checkVocab",
-        entity_or_edge_id: row.id,
-        message: `Edge '${row.id}' has unknown relation_type '${row.relation_type}' (not in RELATION_TYPES registry)`,
-        severity: "warning",
-      });
-    }
-  }
-
-  return violations;
 }
 
 /**
